11.09.2022 Done: 输出默认画面 修改了交换链backbuffer的表示方法 RT不用const 清理RT和切换RT状态 Todo：简单的场景实现

+++ StreEngine/Rendering/include/base_type.h
@@ -122,5 +122,5 @@ struct s_vertex
 };
 
 
-typedef uint32_t s_index;
+typedef uint16_t s_index;
 typedef char s_texture;
